Relocating type files
