+++ src/types/chameleon-platform.d.ts
@@ -11,8 +11,8 @@ import {MysqlConnectionOptions} from 'typeorm/driver/mysql/MysqlConnectionOption
 import * as stream from 'stream';
 import {Terminal} from 'xterm-headless';
 import {SerializeAddon} from 'xterm-addon-serialize';
-import {SocketReceiveMode} from './chameleon-platform.enum';
 import {History} from '../entities/History';
+import {SocketReceiveMode} from "./chameleon-platform.common";
 
 export type Resolver = (value?: unknown) => void;
 type ReadStreamClose = (callback?: (err?: NodeJS.ErrnoException | null) => void) => void;
@@ -76,31 +76,6 @@ export type PlatformConfig = {
     httpPort: number;
     db: MysqlConnectionOptions;
 };
-
-
-export type ModelParameters = {
-    uischema: any;
-    schema: any;
-}
-
-export type ModelConfig = {
-    paths: {
-        script: string;
-        input: string;
-        inputInfo: string;
-        parameters: string;
-        output: string;
-        outputInfo: string;
-        outputDescription: string;
-        controllerDirectory: string;
-        debugLog: string;
-    }
-}
-
-export type TerminalResizeOption = {
-    rows: number;
-    cols: number;
-}
 
 export type SocketHandle = (client: DefaultSocketServer, socket: DefaultSocket, message: any) => void;
 export type WebSocketHandle = (server: DefaultWSServer, socket: DefaultWSocket, data: any) => void;
